Tweaked IO
The .xml no longer saves automode defense data, as it was causing some
issues. (Why do we even have teams save crap anymore?)

A modified copy of the /slightly/ outdated (only 65 rounds) Chaifetz
Arena .xml is added to the root of the project folder for use with this
update.

+++ MyScout/MyScout/src/Classes/Team.cs
@@ -68,8 +68,8 @@ namespace MyScout
         /// </summary>
         public float[] autoDefensesCrossed = new float[9];
         public float autoDefensesReached = 0;
-        public float autoHighGoals = 0;
-        public float autoLowGoals = 0;
+        public int autoHighGoals = 0;
+        public int autoLowGoals = 0;
         #endregion
         #region PreScout
         public bool canScoreHighGoals = false;
